feat: refactor HistoryTable
